+++ tests/PhpSpreadsheetTests/Reader/Xlsx/HiddenWorksheetTest.php
@@ -46,4 +46,45 @@ class HiddenWorksheetTest extends TestCase
             ],
         ];
     }
+
+    public function testListWorksheetInfo(): void
+    {
+        $filename = 'tests/data/Reader/XLSX/visibility.xlsx';
+        $reader = new Xlsx();
+        $expected = [
+            [
+                'worksheetName' => 'Sheet1',
+                'lastColumnLetter' => 'A',
+                'lastColumnIndex' => 0,
+                'totalRows' => 1,
+                'totalColumns' => 1,
+                'sheetState' => 'visible',
+            ],
+            [
+                'worksheetName' => 'Sheet2',
+                'lastColumnLetter' => 'A',
+                'lastColumnIndex' => 0,
+                'totalRows' => 1,
+                'totalColumns' => 1,
+                'sheetState' => 'hidden',
+            ],
+            [
+                'worksheetName' => 'Sheet3',
+                'lastColumnLetter' => 'A',
+                'lastColumnIndex' => 0,
+                'totalRows' => 1,
+                'totalColumns' => 1,
+                'sheetState' => 'visible',
+            ],
+            [
+                'worksheetName' => 'Sheet4',
+                'lastColumnLetter' => 'B',
+                'lastColumnIndex' => 1,
+                'totalRows' => 1,
+                'totalColumns' => 2,
+                'sheetState' => 'veryHidden',
+            ],
+        ];
+        self::assertSame($expected, $reader->listWorksheetInfo($filename));
+    }
 }